add utility for getting git hashes of packages
