+++ src/utils_config/__init__.py
@@ -1,3 +1,4 @@
 from .config_loader import ConfigLoader
+from .version_check import check_imports_and_versions
 
-__all__ = ["ConfigLoader"]
+__all__ = ["ConfigLoader", "check_imports_and_versions"]
